requirements.txt
streamlit==1.45.0
google-genai==1.13.0